Fix/optimize groups
Cicero generation for negative group was broken, before we considered
that "Not Match" was advancing to the next characted, which is not true.
We need to add an addictional "Match Any"

This was found while running Brill benchmark, which uses a lot of
negative groups. A few examples are added to the tests.